move single components to common dir
