+++ src/components/common/Header.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Nav, Navbar, Button } from 'react-bootstrap';
 import { useHistory } from 'react-router-dom';
-import { NavigationLinkInterface } from '../types';
+import { NavigationLinkInterface } from '../../types';
 
 export interface HeaderProps {
   links?: NavigationLinkInterface[];
@@ -11,9 +11,7 @@ export const Header: React.FC<HeaderProps> = ({ links }) => {
   const history = useHistory();
   return (
     <Navbar className="navbar-top" expand="lg">
-      <Navbar.Brand onClick={() => history.push('/')}>
-        Sentiment of Bundestag
-      </Navbar.Brand>
+      <Navbar.Brand onClick={() => history.push('/')}>Sentiment of Bundestag</Navbar.Brand>
       <Navbar.Toggle />
       <Navbar.Collapse>
         <Nav className="mr-auto" />
@@ -21,10 +19,12 @@ export const Header: React.FC<HeaderProps> = ({ links }) => {
           {/*          <Nav.Link>
             Über uns
           </Nav.Link> */}
-          <Nav.Link target="_blank" href="https://github.com/orgs/Sentiments-of-Bundestag/dashboard">
+          <Nav.Link
+            target="_blank"
+            href="https://github.com/orgs/Sentiments-of-Bundestag/dashboard"
+          >
             <Button variant="outline-light">Code</Button>
           </Nav.Link>
-
         </Nav>
       </Navbar.Collapse>
     </Navbar>
