+++ src/views/landing.tsx
@@ -8,7 +8,7 @@ import { getSessionsStart } from '../slices/sessions';
 import { getAllSessions } from '../selectors/sessions';
 
 import AppLayout from '../layout/AppLayout';
-import KpiCard from '../components/cards/KpiCard';
+import KpiCard from '../components/common/KpiCard';
 
 interface LandingProps {}
 
@@ -16,9 +16,7 @@ const Landing: React.FC<LandingProps> = () => {
   const dispatch = useDispatch();
   const history = useHistory();
   const { sessions, areSessionsLoading } = useSelector(getAllSessions);
-  const periodCount = Array.from(
-    new Set(sessions.map((s) => s.legislativePeriod)),
-  ).length;
+  const periodCount = Array.from(new Set(sessions.map((s) => s.legislativePeriod))).length;
 
   useEffect(() => {
     dispatch(getSessionsStart());
@@ -33,11 +31,10 @@ const Landing: React.FC<LandingProps> = () => {
       <AppLayout>
         <h1>Analyse der Interaktionen im Deutschen Bundestag</h1>
         <p className="text-justify">
-          Im Projekt Sentiment of Bundestag hat eine Gruppe von Studenten der
-          HTW Berlin im Wintersemester 2020/21 Plenarprotokolle des Deutschen
-          Bundestags aufbereitet und erforscht. Mithilfe einer Sentiment Analyse
-          wurden die Kommentare und Wortmeldungen der Abgeordneten auf ihre
-          Stimmung hin untersucht. Die Ergebnisse wurden visualisiert und nach
+          Im Projekt Sentiment of Bundestag hat eine Gruppe von Studenten der HTW Berlin im
+          Wintersemester 2020/21 Plenarprotokolle des Deutschen Bundestags aufbereitet und
+          erforscht. Mithilfe einer Sentiment Analyse wurden die Kommentare und Wortmeldungen der
+          Abgeordneten auf ihre Stimmung hin untersucht. Die Ergebnisse wurden visualisiert und nach
           Legislaturperioden sortiert.
         </p>
         <Row>
@@ -57,26 +54,16 @@ const Landing: React.FC<LandingProps> = () => {
           </Col>
         </Row>
         <br />
-        <Button
-          variant="primary"
-          size="lg"
-          block
-          onClick={() => redirectTo('factions')}
-        >
+        <Button variant="primary" size="lg" block onClick={() => redirectTo('factions')}>
           Weiter zu den Parteien
         </Button>
-        <Button
-          variant="secondary"
-          size="lg"
-          block
-          onClick={() => redirectTo('persons')}
-        >
+        <Button variant="secondary" size="lg" block onClick={() => redirectTo('persons')}>
           Weiter zu den Abgeordneten
         </Button>
         <p className="text-justify">
           <br />
-          Die Projektarbeit wurde im Rahmen des Moduls Information Systems unter
-          der Leitung von Prof. Dr. rer. nat. Thomas Hoppe durchgeführt.
+          Die Projektarbeit wurde im Rahmen des Moduls Information Systems unter der Leitung von
+          Prof. Dr. rer. nat. Thomas Hoppe durchgeführt.
         </p>
       </AppLayout>
     </>
